requirements.txt
Pillow==7.0.0
fastapi==0.61.0
Werkzeug==1.0.1
numpy==1.18.2
pytest==6.0.1
pdf2image==1.14.0